# Add simple wasm performance test.