--- a/test/extensions/wasm/test_data/Makefile
+++ b/test/extensions/wasm/test_data/Makefile
@@ -1,6 +1,6 @@
 NO_CONTEXT = true
 
-all: logging_cpp.wasm logging_rust.wasm bad_signature_cpp.wasm segv_cpp.wasm emscripten_cpp.wasm asm2wasm_cpp.wasm stats_cpp.wasm missing_cpp.wasm
+all: logging_cpp.wasm logging_rust.wasm bad_signature_cpp.wasm segv_cpp.wasm emscripten_cpp.wasm asm2wasm_cpp.wasm stats_cpp.wasm missing_cpp.wasm speed_cpp.wasm
 
 include ../../../../api/wasm/cpp/Makefile.base_lite
 include ../../../../api/wasm/rust/Makefile.base