--- a/test/extensions/wasm/test_data/speed_cpp.cc
+++ b/test/extensions/wasm/test_data/speed_cpp.cc
@@ -1,0 +1,11 @@
+// NOLINT(namespace-envoy)
+#include <string>
+
+#include "proxy_wasm_intrinsics.h"
+
+extern "C" EMSCRIPTEN_KEEPALIVE void proxy_onTick(uint32_t /* context_zero */) {
+  uint64_t t;
+  if (WasmResult::Ok != proxy_getCurrentTimeNanoseconds(&t)) {
+    logError("bad result from getCurrentTimeNanoseconds");
+  }
+}