--- a/test/extensions/wasm/wasm_speed_test.cc
+++ b/test/extensions/wasm/wasm_speed_test.cc
@@ -2,7 +2,9 @@
  * Simple WASM speed test.
  *
  * Run with:
- * bazel run -c opt  --linkopt -fuse-ld=lld --define wasm=enabled --config=libc++ wasm_speed_test
+ * TEST_SRCDIR=`pwd` TEST_WORKSPACE=bazel-$(basename `pwd`) bazel run --define wasm=enabled
+ * --config=libc++ -c opt //test/extensions/wasm:wasm_speed_test
+ * Note: "--linkopt -f--fuse-ld" may be required as well depending on the build environment.
  */
 #include <stdio.h>