… test/extensions/wasm/test_data/speed_cpp.wasm