Signed-off-by: John Plevyak <jplevyak@gmail.com>

Optimized results:

Benchmark                                                                Time             CPU       Iterations 
BM_WasmSimpleCallSpeedTest/V8SpeedTest           273 ns          273 ns       2591211
BM_WasmSimpleCallSpeedTest/WavmSpeedTest     221 ns           221 ns      3144155
BM_WasmSimpleCallSpeedTest/NullSpeedTest         107 ns          107 ns       6260650

Net: the cost of calls into and out of the VMs are very small.     WAVM is faster than V8.