@AndrewScheidecker 2019-08-31
Based on some investigation on WAVM's `invoke-bench`, I think `catchRuntimeExceptions` is probably slower than `invokeFunctionUnchecked` in your benchmark. It calls `unwindSignalsAsExceptions`, which calls `Platform::catchSignals`, which was quite slow: I've just made [a commit](https://github.com/WAVM/WAVM/commit/c5e1ee0153651eb5df7c38332691548d9d2f4473) that makes it 10x faster.

FYI you're using a version of the WAVM API from before [a change](https://github.com/WAVM/WAVM/commit/fd3ed14a3ad406080dd3b75bb37752b8274b4ed0) I made recently to merge `invokeFunctionChecked` and `invokeFunctionUnchecked`.

The unified `invokeFunction` is slightly slower than `invokeFunctionUnchecked`, mostly because it calls `unwindSignalsAsExceptions` to ensure that if you have any C++ objects on the stack between the `invokeFunction` and `catchRuntimeExceptions`, that their destructor will be called. The combined overhead of `catchRuntimeExceptions` and `invokeFunction` should be much lower with my `catchSignals` optimization, though.