@jplevyak 2019-09-26
I tried to update our WAVM implementation, but the new version is aborting and/or corrupting memory.  In particular I can build a test case where  std::to_string(7) returns "3454" (or some other bogus result).  I changed the code to use char buf[16]; sprintf(buf, "%d", 7); and it also gets corrupted.  This doesn't happen in V8 and it doesn't happen in the older version of WAVM. 

If you  are curious: https://github.com/jplevyak/envoy-wasm/pull/new/wasm-wavmupgrade

The test //test/extensions/wasm:wasm_test fails.

Any suggestions on how to debug would be appreciated.

@AndrewScheidecker