@AndrewScheidecker 2019-09-26
> I tried to update our WAVM implementation, but the new version is aborting and/or corrupting memory. In particular I can build a test case where std::to_string(7) returns "3454" (or some other bogus result). I changed the code to use char buf[16]; sprintf(buf, "%d", 7); and it also gets corrupted. This doesn't happen in V8 and it doesn't happen in the older version of WAVM.

To be clear, this is for `std::to_string`/`sprintf` in some WASM code hosted by Envoy/WAVM?

Can you reproduce it with a WASM file that works with plain WAVM?

Does it still happen if you use the WAVM commit immediately before this one? https://github.com/WAVM/WAVM/commit/b9d2a112cfd8e9404285dd13f4f1849539bcb5f3