@jplevyak 2019-09-28
Problem solved.  I was not initializing the heap pointer correctly for emscripten.