@PiotrSikora 2019-09-29
@jplevyak could you merge with master and re-run tests to get accurate numbers as of now?