@jplevyak 2019-10-01
Upgraded the version of protobuf which was used with speed_cpp.wasm.